refactor: Inline taxon_workflow
