--- app/models/taxa/taxon_workflow.rb
@@ -1,18 +1,0 @@
-class Taxon < ApplicationRecord
-  include Workflow
-  include Workflow::ExternalTable
-
-  has_one :taxon_state
-
-  workflow do
-    state TaxonState::OLD
-    state TaxonState::WAITING do
-      event :approve, transitions_to: TaxonState::APPROVED
-    end
-    state TaxonState::APPROVED
-  end
-
-  def last_change
-    Change.joins(:versions).where("versions.item_id = ? AND versions.item_type = 'Taxon'", id).last
-  end
-end
